refactor: move single-use utils to usage location
