+++ common/BUILD.bazel
@@ -4,9 +4,7 @@ go_library(
     name = "common",
     srcs = [
         "directives.go",
-        "jq.go",
         "regex.go",
-        "rules.go",
         "set.go",
         "walk.go",
     ],
@@ -16,9 +14,7 @@ go_library(
         "//common/logger",
         "@com_github_emirpasic_gods//sets/treeset",
         "@com_github_emirpasic_gods//utils",
-        "@com_github_itchyny_gojq//:gojq",
         "@gazelle//label",
-        "@gazelle//language",
         "@gazelle//rule",
         "@gazelle//walk",
     ],
