+++ common/bazel/BUILD.bazel
@@ -8,7 +8,5 @@ go_library(
     ],
     importpath = "github.com/aspect-build/aspect-gazelle/common/bazel",
     visibility = ["//visibility:public"],
-    deps = [
-        "//common/bazel/workspace",
-    ],
+    deps = ["//common/bazel/workspace"],
 )
