+++ common/walk.go
@@ -5,31 +5,8 @@ import (
 	"slices"
 	"strings"
 
-	BazelLog "github.com/aspect-build/aspect-gazelle/common/logger"
-	"github.com/bazelbuild/bazel-gazelle/language"
 	"github.com/bazelbuild/bazel-gazelle/walk"
 )
-
-type GazelleWalkFunc func(path string) error
-
-// Walk the directory being generated, respecting any walk generation config.
-func GazelleWalkDir(args language.GenerateArgs, walkFunc GazelleWalkFunc) error {
-	BazelLog.Tracef("GazelleWalkDir: %s", args.Rel)
-
-	// Source files in the primary directory
-	for _, f := range args.RegularFiles {
-		// Skip BUILD files
-		if args.Config.IsValidBuildFileName(f) {
-			continue
-		}
-
-		if walkErr := walkFunc(f); walkErr != nil {
-			return walkErr
-		}
-	}
-
-	return nil
-}
 
 func WalkHasPath(rel, p string) bool {
 	d, err := walk.GetDirInfo(rel)
