+++ language/orion/plugin/plugin.star.go
@@ -10,7 +10,7 @@ import (
 	"maps"
 	"slices"
 
-	starUtils "github.com/aspect-build/aspect-gazelle/common/starlark/utils"
+	starUtils "github.com/aspect-build/aspect-gazelle/language/orion/starlark/utils"
 	"go.starlark.net/starlark"
 )
