+++ language/orion/plugin/target.star.go
@@ -3,7 +3,7 @@ package plugin
 import (
 	"fmt"
 
-	starUtils "github.com/aspect-build/aspect-gazelle/common/starlark/utils"
+	starUtils "github.com/aspect-build/aspect-gazelle/language/orion/starlark/utils"
 
 	"go.starlark.net/starlark"
 )
