+++ language/orion/starlark/stdlib/BUILD.bazel
@@ -3,10 +3,10 @@ load("@rules_go//go:def.bzl", "go_library")
 go_library(
     name = "stdlib",
     srcs = ["path.go"],
-    importpath = "github.com/aspect-build/aspect-gazelle/common/starlark/stdlib",
+    importpath = "github.com/aspect-build/aspect-gazelle/language/orion/starlark/stdlib",
     visibility = ["//visibility:public"],
     deps = [
-        "//common/starlark/utils",
+        "//language/orion/starlark/utils",
         "@net_starlark_go//starlark",
     ],
 )
