+++ language/orion/starlark/stdlib/path.go
@@ -3,7 +3,7 @@ package starlark
 import (
 	"path"
 
-	utils "github.com/aspect-build/aspect-gazelle/common/starlark/utils"
+	utils "github.com/aspect-build/aspect-gazelle/language/orion/starlark/utils"
 
 	"go.starlark.net/starlark"
 )
