+++ language/orion/starlark/utils/BUILD.bazel
@@ -7,7 +7,7 @@ go_library(
         "lib.go",
         "util.go",
     ],
-    importpath = "github.com/aspect-build/aspect-gazelle/common/starlark/utils",
+    importpath = "github.com/aspect-build/aspect-gazelle/language/orion/starlark/utils",
     visibility = ["//visibility:public"],
     deps = [
         "@net_starlark_go//starlark",
