+++ language/orion/starzelle/BUILD.bazel
@@ -10,9 +10,9 @@ go_library(
     visibility = ["//visibility:public"],
     deps = [
         "//common/logger",
-        "//common/starlark",
-        "//common/starlark/utils",
         "//language/orion/plugin",
+        "//language/orion/starlark",
+        "//language/orion/starlark/utils",
         "@com_github_bmatcuk_doublestar_v4//:doublestar",
         "@net_starlark_go//starlark",
     ],
